% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duke_pal.R
\name{duke_pal}
\alias{duke_pal}
\title{Duke Color Palette (Discrete) and Scales}
\usage{
duke_pal()
}
\value{
Character vector of Duke palette HEX codes.
}
\description{
An eight-color colorblind friendly qualitative discrete palette that is
based on colors on the Duke branding guidelines.
}
\examples{
duke_pal()
}
\references{
https://brand.duke.edu/colors/
}
