% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fonts.R
\name{install_atkinson_hyperlegible}
\alias{install_atkinson_hyperlegible}
\title{Install Atkinson Hyperlegible}
\usage{
install_atkinson_hyperlegible()
}
\description{
Atkinson Hyperlegible font is named after Braille Institute founder,
J. Robert Atkinson. What makes it different from traditional typography
design is that it focuses on letterform distinction to increase character
recognition, ultimately improving readability. It is it free for
anyone to use!
}
\note{
\href{http://brailleinstitute.org/wp-content/uploads/2020/11/Atkinson-Hyperlegible-Font-License-2020-1104.pdf}{License}
}
