% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fonts.R
\name{install_eb_garamond}
\alias{install_eb_garamond}
\title{Install Garamond EB}
\usage{
install_eb_garamond()
}
\description{
As a cousin to the Garamond 3 LT used in the Duke wordmark, EB Garamond is a
clean and distinguished serif font that hews closely to the Duke legacy brand.
It is flexible enough to be used in headlines, subheads and body text, but
exercise discretion when using it at smaller font sizes.
}
\note{
\href{https://fonts.google.com/specimen/EB+Garamond/about}{Open Font License}
}
\references{
\href{https://brand.duke.edu/typography/#official}{Duke Brand Typography}
}
