% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_duke_color_discrete.R
\name{scale_duke_color_discrete}
\alias{scale_duke_color_discrete}
\title{Apply discrete color scale}
\usage{
scale_duke_color_discrete(
  ...,
  na.value = "#B5B5B5",
  guide = "legend",
  aesthetics = c("color", "colour")
)
}
\arguments{
\item{...}{Arguments passed on to \code{discrete_scale}.}

\item{na.value}{Color used for NA values}

\item{guide}{the name used to create the guide. See \href{https://www.rdocumentation.org/link/guides?package=ggplot2&version=2.1.0}{\code{guides}} for more info.}

\item{aesthetics}{String or vector of strings detailing what aesthetic features this discrete scale can apply to.}
}
\value{
a visualization with discrete duke color scale
}
\description{
Introduces a Duke branded and accessible discrete color palette to ggplot
geometric objects using color argument. It prioritizes high contrast colors
aligned with Web Content Accessibility Guidelines (WCAG).
}
\details{
Partial code for this function can be attributed to
\href{https://github.com/jrnold/ggthemes/blob/main/R/colorblind.R}{ggthemes}.
}
\examples{
library(ggplot2)
library(palmerpenguins)

# default
ggplot(penguins, aes(x = bill_depth_mm, y = bill_length_mm, color = species)) +
  geom_point()

# vs. with Duke scale
ggplot(penguins, aes(x = bill_depth_mm, y = bill_length_mm, color = species)) +
  geom_point() +
  scale_duke_color_discrete()
}
