% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_duke_continuous.R
\name{scale_duke_continuous}
\alias{scale_duke_continuous}
\title{Apply continuous color scale}
\usage{
scale_duke_continuous(
  ...,
  low = "#E2E6ED",
  high = "#00539B",
  space = "Lab",
  na.value = "#B5B5B5",
  guide = "colourbar",
  aesthetics = c("colour", "color")
)
}
\arguments{
\item{...}{Arguments passed on to \href{http://127.0.0.1:65374/help/library/ggplot2/help/continuous_scale}{continuous_scale}.}

\item{low}{Low end of color gradient.}

\item{high}{High end of color gradient.}

\item{space}{Color space in which to calculate gradient.}

\item{na.value}{Color used for NA values.}

\item{guide}{Type of legend. "colorbar" for continuous scale, "legend" for discrete scale.}

\item{aesthetics}{String or vector of strings detailing what aesthetic features this continuous scale can apply to.}
}
\value{
a visualization with continuous duke color scale
}
\description{
Apply continuous color scale
}
\examples{
# Default example from Hatteras to Duke Royal Blue
plot3 <- ggplot2::ggplot(iris, ggplot2::aes(x=Petal.Length, y = Petal.Width)) + ggplot2::geom_point()
plot3 + scale_duke_continuous()
}
