% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_duke_fill_discrete.R
\name{scale_duke_fill_discrete}
\alias{scale_duke_fill_discrete}
\title{Title}
\usage{
scale_duke_fill_discrete(
  ...,
  na.value = "#B5B5B5",
  guide = "legend",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Arguments passed on to \code{discrete_scale}.}

\item{na.value}{Color used for NA values}

\item{guide}{the name used to create the guide. See \href{https://ggplot2.tidyverse.org/reference/guides.html}{\code{guides()}} for more info.}

\item{aesthetics}{String or vector of strings detailing what aesthetic features this discrete scale can apply to.}
}
\value{
a visualization with discrete duke color scale in fill
}
\description{
Title
}
\examples{
# default
ggplot(penguins, aes(x = species, fill = species)) +
  geom_point()

# vs. with scale_duke_color_discrete()
ggplot(penguins, aes(x = species, fill = species)) +
  geom_bar() +
  scale_duke_fill_discrete()
}
