% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_duke.R
\name{theme_duke}
\alias{theme_duke}
\title{Theme Duke}
\usage{
theme_duke(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{the base size}

\item{base_family}{the base family}

\item{base_line_size}{the baseline size}

\item{base_rect_size}{the base rect}
}
\value{
a plot with Duke colors
}
\description{
Theme Duke
}
\examples{

library(ggplot2)
plot <- ggplot(cars, aes(speed, dist)) +
  geom_point() +
  labs(title = "Duke Blue")

# default theme
plot

# Duke theme
plot + theme_duke()
}
