% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_duke_color_discrete.R
\name{scale_duke_color_discrete}
\alias{scale_duke_color_discrete}
\title{Apply discrete color scale
Partial code for this function can be attributed to \href{https://github.com/jrnold/ggthemes/blob/main/R/colorblind.R}{ggthemes}.}
\usage{
scale_duke_color_discrete(
  ...,
  na.value = "#B5B5B5",
  guide = "legend",
  aesthetics = c("color", "colour"),
  num = 1
)
}
\arguments{
\item{...}{Arguments passed on to \code{discrete_scale}.}

\item{na.value}{Color used for NA values}

\item{guide}{the name used to create the guide. See \href{https://www.rdocumentation.org/link/guides?package=ggplot2&version=2.1.0}{\code{guides}} for more info.}

\item{aesthetics}{String or vector of strings detailing what aesthetic features this discrete scale can apply to.}
}
\value{
a visualization with discrete duke color scale
}
\description{
Apply discrete color scale
Partial code for this function can be attributed to \href{https://github.com/jrnold/ggthemes/blob/main/R/colorblind.R}{ggthemes}.
}
\examples{
#' # Default example
plot3 <- ggplot2::ggplot(iris, ggplot2::aes(x=Petal.Length, y = Petal.Width, color = Species)) + ggplot2::geom_point()
plot3 + scale_duke_discrete()
}
