% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_duke_fill_discrete.R
\name{scale_duke_fill_discrete}
\alias{scale_duke_fill_discrete}
\title{Title}
\usage{
scale_duke_fill_discrete(
  ...,
  na.value = "#B5B5B5",
  guide = "legend",
  aesthetics = "fill",
  num = 1
)
}
\arguments{
\item{...}{Arguments passed on to \code{discrete_scale}.}

\item{na.value}{Color used for NA values}

\item{guide}{the name used to create the guide. See \href{https://www.rdocumentation.org/link/guides?package=ggplot2&version=2.1.0}{\code{guides}} for more info.}

\item{aesthetics}{String or vector of strings detailing what aesthetic features this discrete scale can apply to.}
}
\value{
a visualization with discrete duke color scale in fill
}
\description{
Title
}
\examples{
#' # Default example
plot3 <- ggplot2::ggplot(iris, ggplot2::aes(x=Species, fill = Species)) + ggplot2::geom_bar()
plot3 + scale_duke_discrete()
}
