% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_duke.R
\name{theme_duke}
\alias{theme_duke}
\title{Theme Duke}
\usage{
theme_duke(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{the base size}

\item{base_family}{the base family, AtkinsonHyperlegible by default.}

\item{base_line_size}{the baseline size}

\item{base_rect_size}{the base rect}
}
\value{
a plot with Duke colors
}
\description{
Implements the overall aesthetic and thematic features of the plot.
Striving for a minimalist design, it defines the background, grid line,
text, and legend arguments of the visualization.
}
\details{
This theme aims to use Atkinson Hyperlegible font for text in plots
by default. This font focuses on letterform distinction to increase character
recognition, ultimately improving readability. Currently, the font will be
used if it's installed on the user's system. If not, a message will be
printed directing users to where font can be downloaded and font used in plots
will default to "sans". \code{base_family} can also be set to "EB Garamond", which
is the recommended Duke serif font. If installed on the user's system, this
font can be used in plots with \code{theme_duke(base_family = "EB Garamond")}. If
not, a message will be printed directing users to where font can be downloaded
and font used in plots will default to "serif".
}
\examples{

library(ggplot2)
library(palmerpenguins)

# default
p <- ggplot(penguins, aes(x = bill_depth_mm, y = bill_length_mm, color = species)) +
  geom_point() +
  labs(
    title = "Bill length and depth of penguins",
    subtitle = "Dimensions for Adelie, Chinstrap, and Gentoo Penguins",
    x = "Bill depth (mm)",
    y = "Bill length (mm)",
    color = "Species",
    caption = "Source: palmerpenguins package."
  )
p

# vs. with Duke theme
p +
  theme_duke()

# vs. with Duke theme and scale
p +
  scale_duke_color_discrete() +
  theme_duke()

# with Duke theme, scale, and further customization to theme
p +
  scale_duke_color_discrete() +
  theme_duke() +
  theme(
    plot.title = element_text(color = "red", size = 20),
    plot.background = element_rect(fill = "pink", color = "yellow"),
    panel.grid = element_blank()
  )
}
